Winter Gear Buffs (#2045)
# Description

Gives scarves and the bomber jacket temperature protection so all items
in the Winterdrobe (aside from the novelty xmas crown and Santa hat)
actually function to keep you warm. Scarves give a 20% resistance,
equivalent to winter boots, and the bomber jacket gives a 90%
resistance, equivalent to winter coats and the trench coat.

---

# Changelog

:cl:
- tweak: Gave scarves a 20% temperature resistance
- tweak: Gave the bomber jacket a 90% temperature resistance

---------

Signed-off-by: GNUtopia <93669372+GNUtopia@users.noreply.github.com>