Let Foxes and Dogs Track Scents (#2116)
# Description

I thought it was weird that foxes and dogs can't track scents like
Vulpines, so I fixed it. That's all there is really. I'm just surprised
this wasn't done when Vulpines initially got the ability, it just makes
sense.

This makes it so every dog and fox have the ability to track scents when
controlled by a player. This includes Laika, Ians, Renault, Siobhan,
Waltuh, and Cerberus (since it extends MobCorgi), as well as all foxes
and dogs purchased from cargo.

# Changelog

:cl: PsychoLogik (musicman928)
- tweak: Dogs and Foxes can now track scents (Cerberus included)

---------

Signed-off-by: musicman928 <106891103+musicman928@users.noreply.github.com>