Sunglasses Are More Expensive (#2121)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR changes the sunglasses to be more expensive in loadouts, as I
find the fact that they are as cheap as they are to be very unbalanced.
As per the suggestion of Parky, I have also renamed the default such
that it specifies ones that are rated for flash protection.

---
<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![sunglasses](https://github.com/user-attachments/assets/fafa0734-d6ee-4c9b-a003-6c9c21072c7c)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: sunglasses with flash protection are more expensive in the
loadouts
- tweak: sunglasses are either "default" or "shielded" rather than
"cheap" or "standard." the description has also been changed to reflect
this