The Multiclassing Trait - Restricted Gear (#1317)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds a new trait 'Restricted Gear' that unlocks job requirements for a
curated list of items (most equipment), allowing you to take items from
other jobs at a loadout and trait cost.

The main idea is to allow freedom of choosing items 'initial player
power' at a cost of trait points 'final player power', as well as
effectively automating greytiding. Item selection is moderated through
loadout cost, all the numbers are up for discretion, costs can be
balanced by interpolating the cost of the trait and the cost of
contraband items.

This trait is definitely not suitable for all servers, ~~but it has a
place in lowpop RP downstreams like Grimbly Station which I play on.~~
RIP Grimbly Station

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

This PR only really features yml changes, so it could be considered in a
complete state. However there is also room for design / balance
discussion.

- Should command staff be able to use Restricted Gear? (currently yes)
- Should prisoners be able to use Restricted Gear? (currently no, but
funny)
- - it would also be possible to add item restrictions on restricted
gear so prisoners cant smuggle guns
- Should Restricted Gear grant access to certain items on your current
job? (Heirloom armory weapons for security, lethal .50 for bartender
etc)
- Tweaking of numbers, interpolation between trait cost and loadout
cost.
- ~~Does this trait need a CVar requirement? (there was some discussion
on discord about it)~~
- Probably more that I haven't stated.

Of course, this can also be tailored to any downstream as required.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

Trait name / description:

![image](https://github.com/user-attachments/assets/2043cc3d-c121-472e-9b8c-95e4bc7ca2ab)
(dw death i changed *it's* to *its*)

You can use exchange loadout points to get other jobs' items.

![image](https://github.com/user-attachments/assets/da82fd79-1b47-48dd-8e04-90cd9e3c0697)

Loadout point costs don't apply to your own job.

![image](https://github.com/user-attachments/assets/45b02105-6da7-42a2-b4ca-859a744e907a)


</p>
</details>

---

# Breaking changes

PASSENGERS CAN GET GUNS (provided a minimum playtime requirement).
If you don't want it, comment it out

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added a trait 'Restricted Gear' which unlocks loadout job
restrictions at a significant price.

---------

Signed-off-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>
Co-authored-by: sleepyyapril <flyingkarii@gmail.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>