IPC Brain Fixes (#2123)
# Description

Fixes IPCs having their brains 'wiped' upon being gibbed and allows IPC
brains to be surgically removed and inserted.
(I think what was happening was that you were never considered to be
controlling the posibrain since it wasn't tagged as a brain?) This
incidentally happened as part of me trying to allow you to put MMIs in
IPCs, so look forward to that.

---

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/8e237104-9f2d-402a-a65b-e1052ce6bae1

You can see the lights on the freshly liberated posibrain indicating
that it's active.

</p>
</details>

---

# Changelog

:cl:
- fix: Fixed IPCs losing control of their brain when gibbed
- fix: Fixed IPC brains being unable to be surgically inserted/removed