Anchor Station Fixes + Parallax (#2131)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
Anchor fix PR, with a custom background by pont

not fully sure what all was changed, i was just handed the updated map
file and parallax background images lol

---

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/a0e6db73-6963-43d8-a00d-9e04ea986211

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added custom Parallax to Anchor
- fix: Fixed some bugs with Anchor. Also applies some map fixes