Enhances Roundstart Antagonist Selection to Prevent Roundstart Fail and Allocate All Antags (#1339)
We might never see the end of new antags being added using MakeAntag so
better just cut it off by the root and mitigate the issue by preventing
it from causing the round not to start.

<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

Since [this
issue](https://discord.com/channels/1218698320155906090/1218701887990140999/1314367545226166333)
still persisted to manifest even after an attempt to fix it, this PR
simply prevents the rounds from not starting due to the error by
changing the exception into a log error.

Catching and treating could be done instead, but I don't see many
benefits from that at all, so an error it is.

Also important, after asking a player in ahelp about what settings they
had, I concluded it gets triggered when a player has Thief antagonist
enabled, and have jobs set up in such a way that they remain in lobby
even when they ready up. So they seem "elligible" but they have no
entity at roundstart due to their setting to not get spawned at
roundstart as assistant and wait in lobby when you don't have the jobs
you had selected.

I tried to figure a way to still add the same antagonist it failed to
assign in
[ChooseAntags](https://github.com/Simple-Station/Einstein-Engines/blob/0df6205a88ffcf995e242e9225d6bf8085135b73/Content.Server/Antag/AntagSelectionSystem.cs#L202),
and considered decrementing `i` inside that loop, but then I didn't due
to how decrementing `i` would make the loop prone to stalling. So we
lose one of the antagonist slots, but get the round to start, best
effort.

Cheers.
-----

Edit: Please read [my last
comments](https://github.com/Simple-Station/Einstein-Engines/pull/1339#issuecomment-2558363980).

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->


---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://example.com/thisimageisntreal.png)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Enhances roundstart antagonist selection making sure all slots
are allocated
- fix: Prevents Thief or others antags from causing the round to fail to
start

---------

Signed-off-by: Timfa <timfalken@hotmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: Timfa <timfalken@hotmail.com>