Tau Ceti Foreign Legion Content (#2156)
# Description

This PR adds assets related to the Tau-Ceti Foreign Legion, a faction
that is related to and friendly towards NanoTrasen. Its also integrated
with the Contractors system, where characters who meet the requirement
of being a Biesel Citizen, as well as a War Veteran(served in the TCFL)
can take clothing related to the TCFL.

Security characters who meet these requirements can also replace their
security gear with blue sec armor. Have fun. Oh and I also ported the
Captain Revolver from Aurora, same source as all the others.

# TODO

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/a02db3f9-d69c-4eee-97d9-d42dfb55e991)

</p>
</details>

# Changelog

:cl:
- add: Added loadout drip related to the Tau Ceti Foreign Legion. These
are available to any character that is a Biesel Citizen, and also has
the War Veteran lifepath.
- add: Added a fancy shotgun revolver to the Captain's weapon selection.

---------

Signed-off-by: Timfa <timfalken@hotmail.com>
Co-authored-by: Timfa <timfalken@hotmail.com>