Fixed Penlight UI and Blind Detection (#2139)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Fixed up the penlights again. Made it so that the UI messages stored,
because there was an issue with the UI opening blank since the diagnoses
were already done beforehand, and they weren't readable. Further, I
updated the system to properly detect eye damage and blindess.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Fixed Penlights

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: SleepyScarecrow
- fix: Fixed penlights.