The Hive : Rerotation, AI Satellite, Senior Roles, AI Turrets, Etc. (#2108)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description


---
Adds The Hive back into the rotation. I have included most of what I
could think of to improve it. If I forgot something, please let me know.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![hivesat](https://github.com/user-attachments/assets/7909b4d4-19a2-4cd3-b403-7ac248f3897f)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added The Hive back to rotation.