Lings Can Now Eat Animals (#2167)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Lings can now eat a bunch of animals

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
Died 2025 Born 2025 welcome back rat king 2 rp
Lings can now eat a bunch of animals, now they shouldn't be 100% KOS,
they can co-exist with the crew or do enough rat king-ish shit that they
are deemed KOS despite being able to eat animals.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Om nom nom I love stealing the chef's animals.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->




# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- add: Changelings can now "Co-Exist" with the crew by eating animals /
unwanted pests like xenos.