Lumenblast Shells (#2157)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

adds new Lumenblast shells! Lumenblast shells are a direct upgrade to
beanbags that cause 40 stamina damage and cause a flash on impact. These
will be extremely useful for riots and nonlethal takedowns, as they do
next to no actual damage.

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Lumenblast shells, more powerful beanbag rounds with a flash
effect on impact
- add: Security can now take a sawn off shotgun as a duty weapon; it
comes with 4 lumenblast shells, two already loaded
- tweak: Flares can be loaded into the flare gun

---------

Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>