Shiva/Spider Buffs (#2093)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Allows giant spiders to pull things (so they can have their nests full
of caccoons) and makes it so Shiva injects some nocturine on bite,
allowing her to assist security with takedowns. SHIVA ALSO GETS SECURITY
ACCESS YIPPEEEEEE

Yes, a spider injecting nocturine is scary, but it regens slowly, has a
low max internal storage, and its Shiva.

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- tweak: Shiva now injects a small amount of noc on hit
- tweak: Sec dogs and shiva have both received sec access and the
ability to pull
- tweak: Sec dogs can now use the vulpakin sniff ability

---------

Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>
Co-authored-by: Timfa <timfalken@hotmail.com>