Magboots Changes (#2197)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Removes the epi magboots, which were functionally a reskin, and applies
that sprite to the normal magboots. The old yellow magboots sprite has
been turned into a new item called engineering magboots, which has less
of a speed reduction. New magboots have been added called combat
magboots, which provide the benefits of combat boots and some
resistances, but also are slower than normal magboots.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/e3d02584-4361-422d-ba3f-965a28f50ed4)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Adds Combat Magboots, offering all the benefits of combat boots
and magboots at the cost of a speed penalty
- add: Added engineering magboots, which are magboots with decreased
speed penalties
- tweak: Epi magboots have been merged with normal magboots

---------

Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>