The Syndicate Modsuits PR  (#2162)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# "Raider" and "Jaeger" Syndicate Modsuits

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Syndicate Modsuits, sprites are from TG's files, the "Raider" modsuit is
25 TC, while the "Jaeger" modsuit is 70 TC.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [X] Set up YMLs, make them buyable in the uplink etc.,
- [X] TCJ's renaming and description making or something.
- [x] ODJ's thoughts on the balancing.
---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Example Media Embed](https://i.imgur.com/ABcUjKD.png)
![Example Media Embed](https://imgur.com/Zko1plm.png)
![Example Media Embed](https://imgur.com/ygChnNi.png)
![Example Media Embed](https://imgur.com/Jy55pG0.png)
</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Diggy
- add: "Raider" modsuit is 25 TC, while the "Jaeger" modsuit is 70 TC,
both in your Uplink, BUY NOW
- add: loss

---------

Signed-off-by: BloodfiendishOperator <141253729+Diggy0@users.noreply.github.com>
Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>