Loadout and Gear Changes (#2105)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

A bunch of random miscellaneous changes to gear and loadouts.

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Golemancers can now take utility belts 
- add: Paramedics can now take a portafib, and get advanced meds in belt
- add: Sec can now take a sec gasmask for 4 points, besides Ward/HoS/BSO
who get them for free
- tweak: All military gloves have been given slight resistances, and as
such have been removed from engis loadouts
- tweak: Scalpels can now be put into boots with knife slots
- tweak: Interdyne Herbals are now exclusive to interdyne employees
(until we have a discount system)

---------

Signed-off-by: EctoplasmIsGood <109397347+EctoplasmIsGood@users.noreply.github.com>
Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>