Evil VVizard VVeed (#2179)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
Adds Ousiana dust, allowing one to regenerate their psionic Reroll
without having to wait for Noosphere zap events, at the cost of a bowl,
1 bluespace crystal and 1 normalcy crystal, and fair amount of health.
This allowes for Psions to reliably grow their power throughout a round
so long as Epistemics does their job well and co-ordinates with other
departments for ectoplasm (via either glimmer mobs or Botany) and
offerings to the Oracle.
Note : Only way to get reagent is via crafting.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Make component for returning a psion's Reroll
- [x] Add Ousiana dust reagent
- [x] Add Ousiana dust bowl
- [x] Actually add functional crafting recipe for crafting Ousiana dust
bowl

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>


![image](https://github.com/user-attachments/assets/2799ed85-7c0b-4d32-bef8-7f4b8a895369)


![image](https://github.com/user-attachments/assets/bf77c8a5-a5e2-4cfc-96fc-03815e6fd24e)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Ousiana Dust, Allows one to gain benefit from mind opening drugs
like a Noosphere Zap, crafted with 1 bluespace crystal and 1 normalcy
crystal crushed into a bowl. Probably has no long term psychological
side effects.

---------

Signed-off-by: RocketBoss360 <rocketboss360@gmail.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>