Anoigo Audio Tweak (#2161)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description



Honestly, I'm just throwing anything at the wall to fix the cooldown
issue. I noticed that the power worked but it never played the sound
effect? I'm assuming that's where the issue lies. If it doesn't fix the
cooldown issue, at the very least the sound effect will have a sample
rate consistent with the rest.