(Bandaid Fix for Languages/Accents) I'M LANGUAGING OUT MAN!!! (#2212)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

This PR comments out the CharacterItemGroup for basic languages so you
can take ALL OF THEM (side effects may include using all your trait
slots) and the lines responsible for breaking both languages and accents
(the ones that specify how many slots it takes up in the item group and
how many trait slots it uses)
This is so we can have them functioning until that code gets fixed

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Languages and accents now work. JOHN LANGUAGE IS FINALLY REAL!