Telegnosis Fix (#2209)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Telegnosis currently doesnt let you back into your body, this fixes that
with a mob that actually functions.

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] make telegnosis not RR yourself

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/73b3b6d7-0e89-41d0-93b2-849ecca945fa)


![image](https://github.com/user-attachments/assets/b1dba003-f626-41d4-a31f-d41193a9995e)


![image](https://github.com/user-attachments/assets/c17f9ea5-1d9a-4f97-91a8-9a3e1ea7cbdf)


![image](https://github.com/user-attachments/assets/c3a13d59-d73e-40c0-838f-e92b87c00140)


</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: telegnosis lets you return to your body now

---------

Signed-off-by: RocketBoss360 <rocketboss360@gmail.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>