Add Nanotrasen as an Employer for Every Job (#2220)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds Nanotrasen as an employer for every job, because you are on an NT
station. Also makes it so unemployed people can take clown, mime,
musician, and janitor.

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Nanotrasen employees can take every job again

Co-authored-by: Your Name <EctoplasmIsGood@users.noreply.github.com>