Replace Cybersun with Zavodskoi Interstellar (#2228)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Replaces Cybersun as employer with Zavodskoi Interstellar. 
Originally, I added Cybersun as a stand-in for Zavodskoi, as in the
original lore they didn't exist yet in 2450. However, I've noticed that
having Cybersun only caused confusion, as everyone assumed they were the
Cybersun they knew from other lore.
For that reason, I've decided to fudge year numbers and make Zavodskoi
Interstellar exist. They use the same stamp logo, because that was their
logo and I stole it for Cybersun.

They also get a nice jacket.

---



<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/8b00d9a2-06f3-42d5-ae41-b74e74bc1402)


![image](https://github.com/user-attachments/assets/216aa947-cb97-4304-bef8-6b8ce8eb582e)


![image](https://github.com/user-attachments/assets/5b4f57b9-5344-45ba-b838-c7c3e1dcaa91)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- remove: Removed Cybersun
- add: Added Zavodskoi Interstellar to replace Cybersun (Cybersun was a
stand-in for Zavodskoi, but no longer)