Porting ERT/CC Modsuits (#2200)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Ports over the Responsory (ERT), Inquisitory (ERT Chaplain), Apocryphal
(Deathsquad), Corporate (Centcom Official), and the Praetorian
(Bluesheild) modsuits.

Responsory and Inquisitory have the same stats, with being roughly a ERT
Leader suit, and use normal microreactor power cells, They have the same
bag storage as a ERT bag

Apocryphal and Corporate have the same stats as a deathsquad suit, and
use antique power cells, They have the same storage as a Death Squad bag

Praetorian has the same stats as BSO's hardsuit, with the inclusion of
magboots (Ecto requested this), and the same size bag as a duffel bag


All of these suits are admin spawn only for now.
---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Port Responsory Modsuit
- [x] Port Inquisitory Modsuit
- [x] Port Apocryphal Modsuit
- [x] Port Corporate Modsuit
- [x] Port Praetorian Modsuit
---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/cb8eb55b-5fb9-4d41-808c-66b61e7564f7)

![image](https://github.com/user-attachments/assets/c8f4001a-9aef-4c6d-adb4-f55523cc670f)

(From left to right, Apocryphal, Corporate, Responsory, Inquisitory,
Praetorian)
</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Ported over Responsory, Inquisitory, Apocryphal, Corporate,
Praetorian modsuits from TG station

---------

Signed-off-by: CliveOcelot <165305605+Orion-416@users.noreply.github.com>