Smaller Survival Box (#2169)
# Description

Make survival box smaller. Currently, survival box are large—they occupy
a 3x3 slot in the backpack and have a 4x4 storage space. This PR changes
them to occupy a 2x2 slot in the backpack and have a 3x3 storage space.
The main reason I’m making this PR is that whenever I spawn, an item
drops to the ground. I often don’t notice the dropped item and end up
losing it (Looking at the player spawn areas, there are often items on
the ground most of the time. It doesn’t seem to be just my case.).

---

<details><summary><h1>Media</h1></summary>
<p>

![Captura de tela 2025-04-05
190031](https://github.com/user-attachments/assets/9057876b-42ea-4efb-ac57-c1206c6deba1)

</p>
</details>

---

# Changelog

:cl:
- tweak: Make survival box smaller