[NVS Gax] Minor Fixes (#2249)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

minor fixes on Gax:

* fixed xenoarch doors locking incorrectly
* Named several airlocks, making them easier to locate
* fixed AI turret not connected
* added a mining shuttle console to LO Office

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Emilly
- fix: Gax: minor fixes