Magboots Refactor (#2273)
# Description

Yea so it turns out that magboots were previously extremely dependent on
having other components work with them to do the "Magboot" ability. This
doesn't at all comply with ECS standards, since the expectation I have
is that the MagbootComponent ALONE should be sufficient to make an item
into "Magboots". This PR addresses this issue by refactoring
Magboots(and also ClothingSpeedModifier) to no longer depend on each
other for this behavior. MagbootsComponent is now responsible for
handling its own NoSlip, Gravity Immunity, and Movement Speed. This
behavior will also be extremly useful for supporting things such as
Modsuits and Hardsuit/Tacsuit upgrades, as well as Clothing Attachments
in general(which is another thing I'm working on).

I have attached a video demonstrating that I have tested this PR and
verified that it works.

<details><summary><h1>Media</h1></summary>
<p>


https://github.com/user-attachments/assets/f4b602ff-54b9-4df2-a0a0-e3f691c45bf2

</p>
</details>

# Changelog

:cl:
- fix: Fixed various bugs related to Magboots and Integrated Magboots on
hardsuits.
- tweak: All Modsuits as well as tacsuits contributed by Goobstation now
also include Integrated Magsuits.