Intercomms Fix (#2269)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Description.

---
Intercomms were turbo broken because the frequency didn't change from
the default "common" frequency causing it to be only useful for common.
Now they work for any channel as they were supposed to.

<details><summary><h1>Media</h1></summary>
<p>



https://github.com/user-attachments/assets/3abb1e17-9f1b-4ba4-b165-ba94de82501f


</p>
</details>


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Intercomms are functional for every channel again.